preprocessing/parameter-reverb/pyproject.toml
[project]
name = "parameter-reverb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pedalboard>=0.9.21",
    "python-dotenv>=1.2.1",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "tqdm>=4.67.1",
]
